# Conflicts:
#	features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/detail/presentation/fragment/AddEditProductDetailFragment.kt
#	features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/detail/presentation/viewmodel/AddEditProductDetailViewModel.kt
#	features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/preview/presentation/fragment/AddEditProductPreviewFragment.kt

+++ features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/detail/presentation/viewholder/WholeSaleInputViewHolder.kt
@@ -67,7 +67,7 @@ class WholeSaleInputViewHolder(itemView: View,
                     if (it.isNotEmpty()) {
                         // format the number
                         wholeSalePriceField?.textFieldInput?.removeTextChangedListener(this)
-                        val formattedText: String = NumberFormat.getNumberInstance(Locale.US).format(it.toLong()).toString().replace(",", ".")
+                        val formattedText: String = NumberFormat.getNumberInstance(Locale.US).format(it.toBigInteger()).toString().replace(",", ".")
                         wholeSalePriceField?.textFieldInput?.setText(formattedText)
                         wholeSalePriceField?.textFieldInput?.setSelection(formattedText.length)
                         wholeSalePriceField?.textFieldInput?.addTextChangedListener(this)
