+++ features/merchant/product_manage_common/src/main/java/com/tokopedia/product/manage/common/draft/mapper/AddEditProductDraftMapper.kt
@@ -42,7 +42,7 @@ class AddEditProductDraftMapper {
             if (categoryId.isNotEmpty()) {
                 completionCount++
             }
-            if (price > 0) {
+            if (price > 0.toBigInteger()) {
                 completionCount++
             }
             if (stock > 0) {
